rewritten scrollable side menu
